move BazaarDataManager.ensureConversionsLoaded to ResourceManager
